Struct mirrord_agent::file::FileManager

pub struct FileManager {
    root_path: PathBuf,
    pub open_files: HashMap<u64, RemoteFile>,
    pub dir_streams: HashMap<u64, Enumerate<ReadDir>>,
    pub getdents_streams: HashMap<u64, Peekable<Chain<IntoIter<Result<DirEntryInternal, Error>>, Map<Map<Enumerate<ReadDir>, impl Fn((usize, Result<DirEntry>)) -> Result<DirEntryInternal>>, impl Fn(Result<DirEntryInternal>) -> Result<DirEntryInternal>>>>>,
    index_allocator: IndexAllocator<u64>,
}

Fields
root_path: PathBufopen_files: HashMap<u64, RemoteFile>dir_streams: HashMap<u64, Enumerate<ReadDir>>getdents_streams: HashMap<u64, Peekable<Chain<IntoIter<Result<DirEntryInternal, Error>>, Map<Map<Enumerate<ReadDir>, impl Fn((usize, Result<DirEntry>)) -> Result<DirEntryInternal>>, impl Fn(Result<DirEntryInternal>) -> Result<DirEntryInternal>>>>>index_allocator: IndexAllocator<u64>
Implementations

impl FileManager

pub fn handle_message( &mut self, request: FileRequest ) -> Result<Option<FileResponse>, AgentError>

Executes the request and returns the response.

pub fn new(pid: Option<u64>) -> Self

fn open( &mut self, path: PathBuf, open_options: OpenOptionsInternal ) -> RemoteResult<OpenFileResponse>

fn open_relative( &mut self, relative_fd: u64, path: PathBuf, open_options: OpenOptionsInternal ) -> RemoteResult<OpenFileResponse>

pub(crate) fn read( &mut self, fd: u64, buffer_size: u64 ) -> RemoteResult<ReadFileResponse>

pub(crate) fn read_line( &mut self, fd: u64, buffer_size: u64 ) -> RemoteResult<ReadFileResponse>

Remote implementation of fgets.

Uses BufReader::read_line to read a line (including "\n") from a file with fd. The file cursor position has to be moved manually due to this.

fgets is only supposed to read buffer_size, so we limit moving the file’s position based on it (even though we return the full Vec of bytes).

pub(crate) fn read_limited( &mut self, fd: u64, buffer_size: u64, start_from: u64 ) -> RemoteResult<ReadFileResponse>

pub(crate) fn write_limited( &mut self, fd: u64, start_from: u64, buffer: Vec<u8> ) -> RemoteResult<WriteFileResponse>

pub(crate) fn seek( &mut self, fd: u64, seek_from: SeekFrom ) -> RemoteResult<SeekFileResponse>

pub(crate) fn write( &mut self, fd: u64, write_bytes: Vec<u8> ) -> RemoteResult<WriteFileResponse>

pub(crate) fn close(&mut self, fd: u64)

pub(crate) fn close_dir(&mut self, fd: u64)

pub(crate) fn access( &mut self, pathname: PathBuf, mode: u8 ) -> RemoteResult<AccessFileResponse>

pub(crate) fn xstat( &mut self, path: Option<PathBuf>, fd: Option<u64>, follow_symlink: bool ) -> RemoteResult<XstatResponse>

pub(crate) fn xstatfs(&mut self, fd: u64) -> RemoteResult<XstatFsResponse>

pub(crate) fn fdopen_dir(&mut self, fd: u64) -> RemoteResult<OpenDirResponse>

pub(crate) fn get_dir_stream( &mut self, fd: u64 ) -> RemoteResult<&mut Enumerate<ReadDir>>

fn log_err(entry_res: Result<DirEntryInternal>) -> Result<DirEntryInternal>

fn path_to_dir_entry_internal( path: &Path, position: u64, name: String ) -> Result<DirEntryInternal>

fn get_current_and_parent_entries( current: &Path ) -> IntoIter<Result<DirEntryInternal>>

Get an iterator that contains entries of the current and parent (if exists) directories, to chain with the iterator returned by std::fs::read_dir.

pub(crate) fn get_or_create_getdents64_stream( &mut self, fd: u64 ) -> RemoteResult<&mut Peekable<Chain<IntoIter<Result<DirEntryInternal, Error>>, Map<Map<Enumerate<ReadDir>, impl Fn((usize, Result<DirEntry>)) -> Result<DirEntryInternal>>, impl Fn(Result<DirEntryInternal>) -> Result<DirEntryInternal>>>>>

If a stream does not yet exist for this fd, we create and return it. The possible remote errors are: ResponseError::NotFound if there is not such fd here. ResponseError::NotDirectory if the fd points to a file with a non-directory file type.

pub(crate) fn read_dir(&mut self, fd: u64) -> RemoteResult<ReadDirResponse>

pub(crate) fn getdents64( &mut self, fd: u64, buffer_size: u64 ) -> RemoteResult<GetDEnts64Response>

The getdents64 syscall writes dir entries to a buffer, as long as they fit. If a call did not process all the entries in a dir, the result of the next call continues where the last one stopped. After writing all entries, all future calls return 0 entries. The caller keeps calling until getting 0.