Struct mirrord_agent::outgoing::udp::UdpOutgoingApi

pub(crate) struct UdpOutgoingApi {
    _task: JoinHandle<()>,
    task_status: TaskStatus,
    layer_tx: Sender<LayerUdpOutgoing>,
    daemon_rx: Receiver<DaemonUdpOutgoing>,
}

Handles (briefly) the UdpOutgoingRequest and UdpOutgoingResponse messages, mostly the passing of these messages to the interceptor_task thread.

Fields
_task: JoinHandle<()>

Holds the interceptor_task.

task_status: TaskStatus

Status of the interceptor_task.

layer_tx: Sender<LayerUdpOutgoing>

Sends the Layer message to the interceptor_task.

daemon_rx: Receiver<DaemonUdpOutgoing>

Reads the Daemon message from the interceptor_task.

Implementations

impl UdpOutgoingApi

const TASK_NAME: &'static str = "UdpOutgoing"

pub(crate) fn new(pid: Option<u64>) -> Self

async fn interceptor_task( layer_rx: Receiver<LayerUdpOutgoing>, daemon_tx: Sender<DaemonUdpOutgoing> ) -> Result<(), AgentError>

Does the actual work for Requests and prepares the `Responses:

pub(crate) async fn layer_message( &mut self, message: LayerUdpOutgoing ) -> Result<(), AgentError>

Sends a UdpOutgoingRequest to the interceptor_task.

pub(crate) async fn daemon_message( &mut self ) -> Result<DaemonUdpOutgoing, AgentError>

Receives a UdpOutgoingResponse from the interceptor_task.