Module mirrord_agent::runtime

Modules

crio ğŸ”’

Structs

ContainerInfo ğŸ”’

ContainerdContainer ğŸ”’

DockerContainer ğŸ”’

Enums

Container ğŸ”’

Constants

CONTAINERD_ALTERNATIVE_SOCK_PATH ğŸ”’

CONTAINERD_DEFAULT_SOCK_PATH ğŸ”’

CONTAINERD_K3S_SOCK_PATH ğŸ”’

CONTAINERD_MICROK8S_SOCK_PATH ğŸ”’

CONTAINERD_SOCK_PATHS ğŸ”’

Possible containerd socket paths, evaluated from left to right.

DEFAULT_CONTAINERD_NAMESPACE ğŸ”’

Traits

ContainerRuntime ğŸ”’

Functions

connect ğŸ”’

connect_and_find_container ğŸ”’

Connects to the given containerd socket and returns the client only if the given container exists.

extract_env_from_containerd_spec ğŸ”’

Extract from Spec struct the environment variables as HashMap<K,V>

get_container ğŸ”’

get a container object according to args.

set_namespace