Module mirrord_agent::sniffer

Structs

SnifferCommand ğŸ”’

TCPSession ğŸ”’

TcpConnectionSniffer ğŸ”’

TcpPacketData ğŸ”’

TcpSessionIdentifier ğŸ”’

TcpSnifferApi ğŸ”’

Interface used by clients to interact with the TcpConnectionSniffer. Multiple instances of this struct operate on a single sniffer instance.

Enums

SnifferCommands ğŸ”’

Functions

get_tcp_packet ğŸ”’

is_closed_connection ğŸ”’

is_new_connection ğŸ”’

prepare_sniffer ğŸ”’

resolve_interface ğŸ”’

Connects to a remote address (8.8.8.8:53) so we can find which network interface to use.

Type Aliases

TCPSessionMap ğŸ”’