Struct mirrord_agent::sniffer::TcpConnectionSniffer

pub struct TcpConnectionSniffer {
    port_subscriptions: Subscriptions<Port, u32>,
    receiver: Receiver<SnifferCommand>,
    client_senders: HashMap<u32, Sender<DaemonTcp>>,
    raw_capture: RawCapture,
    sessions: HashMap<TcpSessionIdentifier, TCPSession>,
    connection_id_to_tcp_identifier: HashMap<ConnectionId, TcpSessionIdentifier>,
    index_allocator: IndexAllocator<ConnectionId>,
}

Fields
port_subscriptions: Subscriptions<Port, u32>receiver: Receiver<SnifferCommand>client_senders: HashMap<u32, Sender<DaemonTcp>>raw_capture: RawCapturesessions: HashMap<TcpSessionIdentifier, TCPSession>connection_id_to_tcp_identifier: HashMap<ConnectionId, TcpSessionIdentifier>index_allocator: IndexAllocator<ConnectionId>
Implementations

impl TcpConnectionSniffer

pub const TASK_NAME: &'static str = "Sniffer"

pub async fn start( self, cancel_token: CancellationToken ) -> Result<(), AgentError>

Runs the sniffer loop, capturing packets.

pub async fn new( receiver: Receiver<SnifferCommand>, network_interface: Option<String> ) -> Result<Self, AgentError>

Creates and prepares a new TcpConnectionSniffer that uses BPF filters to capture network packets.

The capture uses a network interface specified by the user, if there is none, then it tries to find a proper one by starting a connection. If this fails, we use “eth0” as a last resort.

fn handle_new_client(&mut self, client_id: u32, sender: Sender<DaemonTcp>)

async fn handle_subscribe( &mut self, client_id: u32, port: Port ) -> Result<(), AgentError>

fn handle_client_closed(&mut self, client_id: u32) -> Result<(), AgentError>

Removes the client with client_id, and also unsubscribes its port.

fn update_sniffer(&mut self) -> Result<(), AgentError>

fn qualified_port(&self, port: u16) -> bool

async fn handle_command( &mut self, command: SnifferCommand ) -> Result<(), AgentError>

async fn send_message_to_clients( &mut self, clients: impl Iterator<Item = &u32>, message: DaemonTcp ) -> Result<(), AgentError>

async fn send_message_to_client( &mut self, client_id: &u32, message: DaemonTcp ) -> Result<(), AgentError>

Sends a DaemonTcp message back to the client with client_id.

async fn handle_packet(&mut self, eth_packet: Vec<u8>) -> Result<(), AgentError>