Struct mirrord_agent::steal::api::TcpStealerApi

pub(crate) struct TcpStealerApi {
    client_id: u32,
    command_tx: Sender<StealerCommand>,
    daemon_rx: Receiver<DaemonTcp>,
    task_status: TaskStatus,
}

Bridges the communication between the agent and the [TcpConnectionStealer] task. There is an API instance for each connected layer (“client”). All API instances send commands On the same stealer command channel, where the layer-independent stealer listens to them.

Fields
client_id: u32

Identifies which layer instance is associated with this API.

command_tx: Sender<StealerCommand>

Channel that allows the agent to communicate with the stealer task.

The agent controls the stealer task through this.

daemon_rx: Receiver<DaemonTcp>

Channel that receives DaemonTcp messages from the stealer worker thread.

This is where we get the messages that should be passed back to agent or layer.

task_status: TaskStatus

View on the stealer task’s status.

Implementations

impl TcpStealerApi

pub(crate) async fn new( client_id: u32, command_tx: Sender<StealerCommand>, task_status: TaskStatus, channel_size: usize ) -> Result<Self, AgentError>

Initializes a TcpStealerApi and sends a message to [TcpConnectionStealer] signaling that we have a new client.

async fn send_command(&mut self, command: Command) -> Result<(), AgentError>

Send command to stealer, with the client id of the client that is using this API instance.

fn try_send_command(&self, command: Command) -> Result<(), AgentError>

Send command synchronously to stealer with try_send, with the client id of the client that is using this API instance.

pub(crate) async fn recv(&mut self) -> Result<DaemonTcp, AgentError>

Helper function that passes the DaemonTcp messages we generated in the [TcpConnectionStealer] task, back to the agent.

Called in the [ClientConnectionHandler].

pub(crate) async fn port_subscribe( &mut self, port_steal: StealType ) -> Result<(), AgentError>

Handles the conversion of LayerTcpSteal::PortSubscribe, that is passed from the agent, to an internal stealer command Command::PortSubscribe.

The actual handling of this message is done in [TcpConnectionStealer].

pub(crate) async fn port_unsubscribe( &mut self, port: Port ) -> Result<(), AgentError>

Handles the conversion of LayerTcpSteal::PortUnsubscribe, that is passed from the agent, to an internal stealer command Command::PortUnsubscribe.

The actual handling of this message is done in [TcpConnectionStealer].

pub(crate) async fn connection_unsubscribe( &mut self, connection_id: ConnectionId ) -> Result<(), AgentError>

Handles the conversion of LayerTcpSteal::ConnectionUnsubscribe, that is passed from the agent, to an internal stealer command Command::ConnectionUnsubscribe.

The actual handling of this message is done in [TcpConnectionStealer].

pub(crate) async fn client_data( &mut self, tcp_data: TcpData ) -> Result<(), AgentError>

Handles the conversion of [LayerTcpSteal::TcpData], that is passed from the agent, to an internal stealer command Command::ResponseData.

The actual handling of this message is done in [TcpConnectionStealer].

pub(crate) async fn http_response( &mut self, response: HttpResponse ) -> Result<(), AgentError>

Handles the conversion of LayerTcpSteal::HttpResponse, that is passed from the agent, to an internal stealer command Command::HttpResponse.

The actual handling of this message is done in [TcpConnectionStealer].

pub(crate) fn close_client(&mut self) -> Result<(), AgentError>

Handles the conversion of [LayerTcpSteal::ClientClose], that is passed from the agent, to an internal stealer command Command::ClientClose.

The actual handling of this message is done in [TcpConnectionStealer].

Called by the Drop implementation of TcpStealerApi.

pub(crate) async fn handle_client_message( &mut self, message: LayerTcpSteal ) -> Result<(), AgentError>