Struct mirrord_agent::steal::connection::TcpConnectionStealer

pub(crate) struct TcpConnectionStealer {

    port_subscriptions: HashMap<Port, StealSubscription>,
    command_rx: Receiver<StealerCommand>,
    clients: HashMap<u32, Sender<DaemonTcp>>,
    index_allocator: IndexAllocator<ConnectionId>,
    stealer: TcpListener,
    iptables: Option<SafeIpTables<IPTables>>,
    write_streams: HashMap<ConnectionId, WriteHalf<TcpStream>>,
    read_streams: StreamMap<ConnectionId, ReaderStream<ReadHalf<TcpStream>>>,
    connection_clients: HashMap<ConnectionId, u32>,
    client_connections: HashMap<u32, HashSet<ConnectionId>>,
    http_request_sender: Sender<HandlerHttpRequest>,
    http_request_receiver: Receiver<HandlerHttpRequest>,
    http_connection_close_sender: Sender<ConnectionId>,
    http_connection_close_receiver: Receiver<ConnectionId>,
    http_connection_clients: HashMap<ConnectionId, HashSet<u32>>,
    http_response_senders: HashMap<(ConnectionId, RequestId), Sender<Response<Full<Bytes>>>>,

}

Created once per agent during initialization.

Runs as a separate thread while the agent lives.

(agent -> stealer) communication is handled by [command_rx];

(stealer -> agent) communication is handled by [client_senders], and the Sender channels come inside StealerCommands through [command_rx];

Fields
port_subscriptions: HashMap<Port, StealSubscription>

Maps a port to its active subscription.

command_rx: Receiver<StealerCommand>

Communication between (agent -> stealer) task.

The agent controls the stealer task through [TcpStealerAPI::command_tx].

clients: HashMap<u32, Sender<DaemonTcp>>

Connected clients (layer instances) and the channels which the stealer task uses to send back messages (stealer -> agent -> layer).

index_allocator: IndexAllocator<ConnectionId>stealer: TcpListener

Intercepts the connections, instead of letting them go through their normal pathways, this is used to steal the traffic.

iptables: Option<SafeIpTables<IPTables>>

Set of rules the agent uses to steal traffic from through the TcpConnectionStealer::stealer listener. None when there are no subscribers.

write_streams: HashMap<ConnectionId, WriteHalf<TcpStream>>

Used to send data back to the original remote connection.

read_streams: StreamMap<ConnectionId, ReaderStream<ReadHalf<TcpStream>>>

Used to read data from the remote connections.

connection_clients: HashMap<ConnectionId, u32>

Associates a ConnectionId with a ClientID, so we can send the data we read from TcpConnectionStealer::read_streams to the appropriate client (layer).

client_connections: HashMap<u32, HashSet<ConnectionId>>

Map a ClientId to a set of its ConnectionIds. Used to close all connections when client closes.

http_request_sender: Sender<HandlerHttpRequest>

Mspc sender to clone and give http filter managers so that they can send back requests.

http_request_receiver: Receiver<HandlerHttpRequest>

Receives filtered HTTP requests that need to be forwarded a client.

http_connection_close_sender: Sender<ConnectionId>

For informing the Self::start task about closed connections.

http_connection_close_receiver: Receiver<ConnectionId>

Self::start listens to this, removes the connection and frees the index.

http_connection_clients: HashMap<ConnectionId, HashSet<u32>>

Keep track of the clients that already received an http request out of a connection. This is used to inform them when the connection is closed.

Note: The set of clients here is not the same as the set of clients that subscribe to the port of a connection, as some clients might have defined a regex that no request matched, so they did not get any request out of this connection, so they are not even aware of this connection.

http_response_senders: HashMap<(ConnectionId, RequestId), Sender<Response<Full<Bytes>>>>

Maps each pending request id to the sender into the channel with the hyper service that received that requests and is waiting for the response.

Implementations

impl TcpConnectionStealer

pub const TASK_NAME: &'static str = "Stealer"

pub(crate) async fn new( command_rx: Receiver<StealerCommand> ) -> Result<Self, AgentError>

Initializes a new TcpConnectionStealer fields, but doesn’t start the actual working task (call TcpConnectionStealer::start to do so).

fn iptables(&self) -> Result<&SafeIpTables<IPTables>, AgentError>

Get a result with a reference to the iptables. Should only be called while there are subscribers (otherwise self.iptables is None).

pub(crate) async fn start( self, cancellation_token: CancellationToken ) -> Result<(), AgentError>

Runs the tcp traffic stealer loop.

The loop deals with 6 different paths:

Receiving StealerCommands and calling TcpConnectionStealer::handle_command;

Accepting remote connections through the TcpConnectionStealer::stealer TcpListener. We steal traffic from the created streams.

Reading incoming data from the stolen remote connections (accepted in 2.) and forwarding to clients.

Receiving filtered HTTP requests and forwarding them to clients (layers).

Receiving the connection IDs of closing filtered HTTP connections, and informing all clients that were forward a request out of that connection of the closing of that connection.

Handling the cancellation of the whole stealer thread.

async fn forward_stolen_http_request( &mut self, request: Option<HandlerHttpRequest> ) -> Result<(), AgentError>

Forward a stolen HTTP request from the http filter to the direction of the layer.

HttpFilter –> Stealer –> Layer –> Local App ^- You are here.

async fn forward_incoming_tcp_data( &mut self, connection_id: ConnectionId, incoming_data: Option<Result<Bytes, Error>> ) -> Result<(), AgentError>

Forwards data from a remote stream to the client with connection_id.

async fn steal_connection( &mut self, client_id: u32, address: SocketAddr, port: Port, stream: TcpStream ) -> Result<(), AgentError>

Forward the whole connection to given client.

async fn incoming_connection( &mut self, __arg1: (TcpStream, SocketAddr) ) -> Result<(), AgentError>

Handles a new remote connection that was accepted on the TcpConnectionStealer::stealer listener.

We separate the stream created by accepting the connection into ReadHalf and WriteHalf to handle reading and sending separately.

Also creates an association between connection_id and client_id to be used by [forward_incoming_tcp_data].

fn new_client(&mut self, client_id: u32, sender: Sender<DaemonTcp>)

Registers a new layer instance that has the steal feature enabled.

async fn init_iptables(&mut self) -> Result<(), AgentError>

Initialize iptables member, which creates an iptables chain for our rules.

async fn port_subscribe( &mut self, client_id: u32, port_steal: StealType ) -> Result<(), AgentError>

Helper function to handle Command::PortSubscribe messages.

Inserts port into TcpConnectionStealer::iptables rules, and subscribes the layer with client_id to steal traffic from it.

async fn port_unsubscribe( &mut self, client_id: u32, port: Port ) -> Result<(), AgentError>

Helper function to handle Command::PortUnsubscribe messages.

Removes port from TcpConnectionStealer::iptables rules, and unsubscribes the layer with client_id.

fn get_client_ports(&self, client_id: u32) -> Vec<Port>

async fn close_client(&mut self, client_id: u32) -> Result<(), AgentError>

Removes the client with client_id from our list of clients (layers), and also removes their redirection rules from TcpConnectionStealer::iptables and all their open connections.

async fn send_message_to_single_client( &mut self, client_id: &u32, message: DaemonTcp ) -> Result<(), AgentError>

Sends a DaemonTcp message back to the client with client_id.

async fn forward_data(&mut self, tcp_data: TcpData) -> Result<(), AgentError>

Write the data received from local app via layer to the stream with end client.

async fn http_response(&mut self, response: HttpResponse)

Forward an HTTP response to a stolen HTTP request from the layer back to the HTTP client.

                    _______________agent_______________

Local App –> Layer –> ClientConnectionHandler –> Stealer –> Browser ^- You are here.

fn remove_connection(&mut self, connection_id: ConnectionId) -> Option<u32>

Removes the (ReadHalf, WriteHalf) pair of streams, disconnecting the remote connection. Also remove connection from connection mappings and free the connection index. This method does not remove from client_connections so that it can be called while

fn connection_unsubscribe(&mut self, connection_id: ConnectionId)

Close the connection, remove the id from all maps and free the id.

async fn handle_command( &mut self, command: StealerCommand ) -> Result<(), AgentError>

Handles Commands that were received by TcpConnectionStealer::command_rx.