Module mirrord_agent::steal::http::filter

Contains a couple of handy constants and the filter_task.

filter_task

The meaty part of our HTTP traffic stealing feature, that creates the whole HTTP filtering mechanism.

TokioExecutor

Temporary affair required for HTTP/2 handshaking in hyper.

close_connection

Notifies that the connection for this current filter_task should be closed.

Structs

HttpHeaderFilter 🔒

HttpPathFilter 🔒

RequestMatchCandidate 🔒

Struct to cache/hold/reuse the attributes that a Request can be matched upon

TokioExecutor

Future executor that utilises tokio threads.

Enums

HttpFilter 🔒

Current supported filtering criterias.

Constants

DEFAULT_HTTP_VERSION_DETECTION_TIMEOUT 🔒

Timeout value for how long we wait for a stream to contain enough bytes to assert which HTTP version we’re dealing with.

H2_PREFACE 🔒

Default start of an HTTP/2 request.

MINIMAL_HEADER_SIZE 🔒

Controls the amount of data we read when trying to detect if the stream’s first message contains an HTTP request.

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Traits

RequestMatch 🔒

Trait that needs to be implemented for each HttpFilter variant.

Functions

close_connection 🔒

Notifies the filter_task that the connection with connection_id should be closed.

filter_task 🔒

Reads the start of the TcpStream, and decides if it’s HTTP (we currently only support HTTP/1) or not,