Function mirrord_agent::steal::http::hyper_handler::unmatched_request

async fn unmatched_request<const IS_UPGRADE: bool>(
    request: Request<Incoming>,
    upgrade_tx: Option<Sender<RawHyperConnection>>,
    original_destination: SocketAddr
) -> Result<Response<Full<Bytes>>, HttpTrafficError>

Handles the case when no filter matches a header in the request (or we have an HTTP upgrade request).

Flow

Creates a http1::Connection to the original_destination;

Sends the Request to it, and awaits a Response;

Sends the [HttpResponse] back on the connected TcpStream;

Special case (HTTP upgrade request)

If the Request is an HTTP upgrade request, then we send the original_destination connection, through upgrade_tx, to be handled in filter_task.

Why not use hyper::upgrade::on?

hyper::upgrade::on requires the original Request as a parameter, due to it having the OnUpgrade receiver tucked inside as an Extensions (you can see this here).

hyper::upgrade::on polls this receiver to identify if this an HTTP upgrade request.

The issue for us is that we need to send the Request to its original destination, with SendRequest, which takes ownership of the request, prohibiting us to also pass it to the proper hyper upgrade handler.

Trying to copy the Request is futile, as we can’t copy the OnUpgrade extension, and if we move it from the original Request to a copied Request, the channel will never receive anything due it being in a different Request than the one we actually send to the hyper machine.