Struct mirrord_agent::steal::http::hyper_handler::HyperHandler

pub(super) struct HyperHandler<V>where
    V: HttpV + Debug,{
    pub(super) filters: Arc<DashMap<u32, Regex>>,
    pub(super) matched_tx: Sender<HandlerHttpRequest>,
    pub(crate) connection_id: ConnectionId,
    pub(crate) port: Port,
    pub(crate) original_destination: SocketAddr,
    pub(crate) request_id: RequestId,
    pub(super) handle_version: V,
}

Used to pass data to the [Service] implementation of hyper.

Think of this struct as if it was a bunch of function arguments that are passed down to [Service::call] method.

Each TcpStream connection (for stealer) requires this.

Fields
filters: Arc<DashMap<u32, Regex>>

The (shared with the stealer) HTTP filter regexes that are used to filter traffic for this particular connection.

matched_tx: Sender<HandlerHttpRequest>

Sender part of the channel used to communicate with the agent that we have a MatchedHttpRequest, and it should be forwarded to the layer.

connection_id: ConnectionId

Identifies this TcpStream connection.

port: Port

The port we’re filtering HTTP traffic on.

original_destination: SocketAddr

The original SocketAddr of the connection we’re intercepting.

Used for the case where we have an unmatched request (HTTP request did not match any of the filters).

request_id: RequestId

Keeps track of which HTTP request we’re dealing with, so we don’t mix up Requests.

handle_version: V
Implementations

impl<V> HyperHandler<V>where V: HttpV + Debug,

async fn unmatched_request( request: Request<Incoming>, upgrade_tx: Option<Sender<RawHyperConnection>>, original_destination: SocketAddr ) -> Result<Response<Full<Bytes>>, HttpTrafficError>

Handles the case when no filter matches a header in the request (or we have an HTTP upgrade request).

HTTP/1

Flow

Creates a http1::Connection to the original_destination;

Sends the Request to it, and awaits a Response;

Sends the [HttpResponse] back on the connected TcpStream;

Special case (HTTP upgrade request)

If the Request is an HTTP upgrade request, then we send the original_destination connection, through upgrade_tx, to be handled in filter_task.

Why not use hyper::upgrade::on?

hyper::upgrade::on requires the original Request as a parameter, due to it having the OnUpgrade receiver tucked inside as an Extensions (you can see this here).

hyper::upgrade::on polls this receiver to identify if this an HTTP upgrade request.

The issue for us is that we need to send the Request to its original destination, with SendRequest, which takes ownership of the request, prohibiting us to also pass it to the proper hyper upgrade handler.

Trying to copy the Request is futile, as we can’t copy the OnUpgrade extension, and if we move it from the original Request to a copied Request, the channel will never receive anything due it being in a different Request than the one we actually send to the hyper machine.

HTTP/2

Flow

Creates a http2::Connection to the original_destination;

Sends the Request to it, and awaits a Response;

Sends the [HttpResponse] back on the connected TcpStream;

pub(super) async fn handle_request( request: Request<Incoming>, original_destination: SocketAddr, upgrade_tx: Option<Sender<RawHyperConnection>>, filters: Arc<DashMap<u32, Regex>>, port: Port, connection_id: ConnectionId, request_id: RequestId, matched_tx: Sender<HandlerHttpRequest> ) -> Result<Response<Full<Bytes>>, HttpTrafficError>

Handles the incoming HTTP/V Request.

Checks if this Request contains an upgrade header, and if not, then checks if a header matches one of the user specified filters.

Helper function due to the fact that [Service::call] is not an async function.

impl HyperHandler<HttpV1>

pub(super) fn new( filters: Arc<DashMap<u32, Regex>>, matched_tx: Sender<HandlerHttpRequest>, connection_id: ConnectionId, port: Port, original_destination: SocketAddr, upgrade_tx: Option<Sender<RawHyperConnection>> ) -> Self

Creates a new HyperHandler specifically tuned to handle HTTP/1 requests.

impl HyperHandler<HttpV2>

pub(crate) fn new( filters: Arc<DashMap<u32, Regex>>, matched_tx: Sender<HandlerHttpRequest>, connection_id: ConnectionId, port: Port, original_destination: SocketAddr ) -> Self

Creates a new HyperHandler specifically tuned to handle HTTP/2 requests.