Trait mirrord_agent::steal::http::HttpV

trait HttpV {
    type Sender;

    // Required methods
    async fn serve_connection(
        stream: ReversibleStream<MINIMAL_HEADER_SIZE>,
        original_destination: SocketAddr,
        connection_id: ConnectionId,
        filters: Arc<DashMap<u32, Regex>>,
        matched_tx: Sender<HandlerHttpRequest>,
        connection_close_sender: Sender<ConnectionId>
    ) -> Result<(), HttpTrafficError>;
    async fn connect(
        target_stream: TcpStream,
        upgrade_tx: Option<Sender<RawHyperConnection>>
    ) -> Result<Self::Sender, HttpTrafficError>;
    async fn send_request(
        sender: &mut Self::Sender,
        request: Request<Incoming>
    ) -> Result<Response<Full<Bytes>>, HttpTrafficError>;
    fn is_upgrade(r: &Request<Incoming>) -> bool;
}

Unifies hyper handling for HTTP/1 and HTTP/2.

Details

As most of the hyper types around HTTP/1 and HTTP/2 are different, and do not share a trait, we use HttpV to create a shared implementation that is used by hyper_handler::HyperHandler.

Required Associated Types

type Sender

Type for hyper’s SendRequest.

It’s a different type for HTTP/1 (http1::SendRequest) and HTTP/2 (http2::SendRequest).

Required Methods

async fn serve_connection( stream: ReversibleStream<MINIMAL_HEADER_SIZE>, original_destination: SocketAddr, connection_id: ConnectionId, filters: Arc<DashMap<u32, Regex>>, matched_tx: Sender<HandlerHttpRequest>, connection_close_sender: Sender<ConnectionId> ) -> Result<(), HttpTrafficError>

Handles the HTTP connection accordingly to its version.

See filter_task for more details.

async fn connect( target_stream: TcpStream, upgrade_tx: Option<Sender<RawHyperConnection>> ) -> Result<Self::Sender, HttpTrafficError>

Performs a client handshake with target_stream, creating an HTTP connection.

HTTP/1

The HTTP/1 connection is a bit more involved, as we have to deal with potential UPGRADE requests.

We do this manually, by keeping the connection alive with poll_without_shutdown, and by sending it as a RawHyperConnection through upgrade_tx to the HTTP stealer handler.

async fn send_request( sender: &mut Self::Sender, request: Request<Incoming> ) -> Result<Response<Full<Bytes>>, HttpTrafficError>

Sends the request to the original destination with HttpV::Sender.

fn is_upgrade(r: &Request<Incoming>) -> bool

Returns true if this Request contains an UPGRADE header.

Warning:

This implementation should always return false for HTTP/2.