Module mirrord_agent::steal::ip_tables

Modules

chain 🔒

flush_connections 🔒

Flush connections - feature that enables the agent to steal connections that are in progress What do you mean? Imagine there was an ongoing session between Client and Server, then the agent starts, the layer asks to listen on the same port as the Server, and then the agent will only get the next connection since the redirection happens on the nat table which is hit only for new connections. Flush connections overcomes this by marking all existing connections of a specific port, and adding a rule that marked connections will be rejected.

mesh 🔒

redirect 🔒

standard 🔒

Structs

IPTablesWrapper

SafeIpTables 🔒

Wrapper struct for IPTables so it flushes on drop.

Enums

Redirects 🔒

Constants

IPTABLES_TABLE_NAME 🔒

Statics

IPTABLE_INPUT

IPTABLE_INPUT_ENV

IPTABLE_MESH 🔒

IPTABLE_MESH_ENV 🔒

IPTABLE_PREROUTING

IPTABLE_PREROUTING_ENV

IPTABLE_STANDARD 🔒

IPTABLE_STANDARD_ENV 🔒

Traits

IPTables 🔒