Function mirrord_agent::steal::http::filter::filter_task

pub(super) async fn filter_task(
    stolen_stream: TcpStream,
    original_destination: SocketAddr,
    connection_id: ConnectionId,
    filters: Arc<DashMap<u32, Regex>>,
    matched_tx: Sender<HandlerHttpRequest>,
    connection_close_sender: Sender<ConnectionId>
) -> Result<(), HttpTrafficError>

Reads the start of the TcpStream, and decides if it’s HTTP (we currently only support HTTP/1) or not,

HTTP/1

If the stream is identified as HTTP/1 by our check in HttpVersion::new, then we serve the connection with HyperHandler.

Upgrade

If an upgrade request is detected in the HyperHandler, then we take the HTTP connection that’s being served (after HTTP processing is done), and use copy_bidirectional to copy the data from the upgraded connection to its original destination (similar to the Not HTTP/1 handling).

Not HTTP/1

Forwards the whole TCP connection to the original destination with copy_bidirectional.

It’s important to note that, we don’t lose the bytes read from the original stream, due to us converting it into a DefaultReversibleStream.