Module mirrord_agent::steal::http::hyper_handler

The core part of the whole HTTP traffic stealing feature.

HyperHandler

RawHyperConnection

Structs

HyperHandler 🔒

Used to pass data to the [Service] implementation of hyper.

RawHyperConnection 🔒

Holds a connection and bytes that were left unprocessed by the hyper state machine.

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Functions

collect_response 🔒

Converts the body of this response from Incoming into Full.

match_request 🔒

Checks if the Request’s matches any of the filters.

matched_request 🔒

Sends a MatchedHttpRequest through tx to be handled by the stealer -> layer, and then waits for the response and returns it once it’s there.

prepare_response 🔒

Remove headers that would be invalid due to us fiddling with the body, and rebuilds the Response.