Struct mirrord_agent::State

pub(crate) struct State {
    pub(crate) clients: HashSet<u32>,
    pub(crate) index_allocator: IndexAllocator<u32>,
    pub(crate) should_pause: bool,
    pub(crate) container: Option<Container>,
}

Keeps track of connected clients. If pausing target, also pauses and unpauses when number of clients changes from or to 0.

Fields
clients: HashSet<u32>index_allocator: IndexAllocator<u32>should_pause: bool

Was the pause argument passed? If true, will pause the container when no clients are connected.

container: Option<Container>

This is an option because it is acceptable not to pass a container runtime and id if not pausing. When those args are not passed, container is None.

Implementations

impl State

pub async fn new(args: &Args) -> Result<State, AgentError>

Returns Err if container runtime operations failed or if the pause arg was passed, but the container info (runtime and id) was not.

pub async fn get_container_pid(&self) -> Result<Option<u64>, AgentError>

Get the external pid of the target container, if container info available.

pub async fn new_client(&mut self) -> Result<u32, AgentError>

If there are clientIDs left, insert new one and return it. If there were no clients before, and there is a Pauser, start pausing. Propagate container runtime errors.

pub async fn new_connection(
&mut self,
stream: TcpStream,
sniffer_command_tx: Sender<SnifferCommand>,
stealer_command_tx: Sender<StealerCommand>,
cancellation_token: CancellationToken,
dns_sender: Sender<DnsRequest>,
ephemeral_container: bool,
pid: Option<u64>
) -> Result<Option<JoinHandle<u32>>, AgentError>

pub(crate) fn generate_id(&mut self) -> Option<u32>

pub async fn remove_client(&mut self, client_id: u32) -> Result<(), AgentError>

If that was the last client and we are pausing, stop pausing. Propagate container runtime errors.

pub fn no_clients_left(&self) -> bool